Renamed stage 3 to volume processing
