+++ config/algorithm_lists/seaice_stage_2/seaice_stage_2_A001.yml
@@ -14,4 +14,4 @@ algorithms:
   - alg_sla_calculations
   - alg_fbd_calculations
   - alg_thk_calculations
-  - alg_grid_for_volume
+  - alg_merge_months
