+++ config/chain_configs/seaice_stage_2/seaice_stage_2_A001.yml
@@ -66,8 +66,8 @@ alg_thk_calculations:
   rho_sea: 1023.9
   
 
-
-alg_grid_for_volume:
-  grid_directory: ${HOME}/sea_ice_processor/grid_files/
+alg_merge_months:
+  merge_file_dir: ${HOME}/sea_ice_processor/month_files/
+  mp_file_timeout: 30
 
 # Resource Locators
