+++ src/clev2er/algorithms/seaice_stage_3_volume/tests/test_alg_vol_fill_nn.py
@@ -1,5 +1,5 @@
 """pytest for algorithm
-clev2er.algorithms.seaice_stage_3.alg_add_ocean_frac
+clev2er.algorithms.seaice_stage_3_volume.alg_vol_totals.py
 """
 
 import logging
@@ -11,7 +11,10 @@ import numpy as np
 import pytest
 from netCDF4 import Dataset  # pylint:disable=no-name-in-module
 
-from clev2er.algorithms.seaice_stage_3.alg_add_ocean_frac import Algorithm
+from clev2er.algorithms.seaice_stage_3_volume.alg_vol_calculations import (
+    Algorithm as CalcVol,
+)
+from clev2er.algorithms.seaice_stage_3_volume.alg_vol_fill_nn import Algorithm
 from clev2er.utils.config.load_config_settings import load_config_files
 
 logger = logging.getLogger(__name__)
@@ -48,7 +51,9 @@ def previous_steps(
     """
     ## Initialise the previous chain steps (needed to test current step properly)
     try:
-        chain_previous_steps: Dict[str, Any] = {}
+        chain_previous_steps: Dict[str, Any] = {
+            "vol_calculations": CalcVol(config, logger),
+        }
     except KeyError as exc:
         raise RuntimeError(f"Could not initialize previous steps in chain {exc}") from exc
 
@@ -67,65 +72,60 @@ def thisalg(config: Dict) -> Algorithm:  # pylint: disable=redefined-outer-name
     """
     # Initialise the Algorithm
     try:
-        this_algo = Algorithm(config, logger)  # no config used for this alg
+        this_algo = Algorithm(config, logger)
     except KeyError as exc:
         raise RuntimeError(f"Could not initialize algorithm {exc}") from exc
 
     return this_algo
 
 
-merge_file_test = [(0), (1)]
-
-
-@pytest.mark.parametrize("file_num", merge_file_test)
-def test_add_ocean_frac_sar(
-    file_num,
+def test_vol_total(
     previous_steps: Dict,
     thisalg: Algorithm,  # pylint: disable=redefined-outer-name
 ) -> None:
-    """test alg_add_ocean_frac.py
+    """test alg_vol_total.py
 
     Test plan:
-    Load a merge file
+    Load a grid file
     run Algorithm.process() on each
     test that the files return (True, "")
-    test that 'ocean_frac' is in shared_dict, it is an array of floats
+    test that volume_grid has been changed using the nn algorithm
     """
 
     base_dir = Path(os.environ["CLEV2ER_BASE_DIR"])
     assert base_dir is not None
 
-    # ================================== MERGE FILE TESTING ========================================
-    logger.info("Testing merge file:")
+    # ==================================  FILE TESTING ==========================================
+    logger.info("Testing  file:")
 
-    # load merge file
-    l1b_merge_file = list(
-        (base_dir / "testdata" / "cs2" / "l1bfiles" / "arctic" / "merge_modes").glob("*.nc")
-    )[file_num]
+    # load  file
+    grid_file = list(
+        (base_dir / "testdata" / "cs2" / "l1bfiles" / "arctic" / "grid_files").glob("*.nc")
+    )[0]
 
     try:
-        l1b = Dataset(l1b_merge_file)
-        logger.info("Loaded %s", l1b_merge_file)
+        l1b = Dataset(grid_file)
+        logger.info("Loaded %s", grid_file)
     except IOError:
-        assert False, f"{l1b_merge_file} could not be read"
+        assert False, f"{grid_file} could not be read"
 
     shared_dict: Dict[str, Any] = {}
 
     for title, step in previous_steps.items():
         success, err_str = step.process(l1b, shared_dict)  # type: ignore[attr-defined]
         if not success:
-            logger.error(" Error with previous step: %s\n%s", title, err_str)
+            logger.error("Error with previous step: %s\n%s", title, err_str)
+
+    grid_before_fill = shared_dict["volume_grid"]
 
     success, err_str = thisalg.process(l1b, shared_dict)
 
-    assert success, f" Algorithm failed due to: {err_str}"
+    assert success, f"Algorithm failed due to: {err_str}"
 
     # Algorithm tests
-    assert "ocean_frac" in shared_dict, "'ocean_frac' not in shared_dict."
+    grid_after_fill = shared_dict["volume_grid"]
 
-    assert isinstance(
-        shared_dict["ocean_frac"], np.ndarray
-    ), f"'ocean_frac' is {type(shared_dict['ocean_frac'])}, not ndarray."
-
-    elev_dtype = str(shared_dict["ocean_frac"].dtype)
-    assert "float" in elev_dtype.lower(), f"Dtype of 'ocean_frac' is {elev_dtype}, not float."
+    # Doesn't always mean that the algorithm didn't work, can improve with:
+    # - Finding a test file that definitely had empty values to fill
+    # - Finding a way of testing if a grid has places that are empty and can be filled using NNs
+    assert not np.array_equal(grid_before_fill, grid_after_fill), "Algorithm had no effect on grid"
